[Refactor/#147]: skeleton 폴더 정리
