+++ src/components/Fallback/Skeleton/Book/BookInfoSkeleton.tsx
